% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delgosha_package.R
\name{delgosha-exports}
\alias{delgosha-exports}
\alias{\%>\%}
\title{hrbrthemes exported operators}
\description{
The following functions are imported and then re-exported
from the hrbrthemes package to enable use of the magrittr
pipe operator with no additional library calls
}
