% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delgosha.R
\docType{package}
\name{delgosha}
\alias{delgosha}
\alias{delgosha-package}
\title{Additional 'ggplot2' Themes and Theme Components designed for RTL languages and specifically for Persian}
\description{
A compilation of extra themes and theme components for 'ggplot2' with an
emphasis on typography.
}
\author{
Muhammad Chenariya Nakhaee (mcnakhaee.com)
}
\keyword{internal}
