% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.R
\name{import_hamshahri}
\alias{import_hamshahri}
\title{Import "Hamshahri" font (deprecated)}
\usage{
import_hamshahri()
}
\description{
Imports Hamshahri
}
\details{
import_*() functions taken from hrbrthemes.
}
\seealso{
\code{\link[extrafont]{font_import}}
}
