% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fa.R
\name{theme_fa}
\alias{theme_fa}
\title{transform any ggplot2 theme to an RTL theme with Persian fonts}
\usage{
theme_fa(
  base_font = "Nahid FD",
  base_size = 12.5,
  plot_title_font = "Nahid FD",
  plot_title_size = 30,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_font = base_font,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_font = base_font,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_font = base_font,
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  legend_font = "Nahid FD",
  legend_font_size = 10,
  legend_color = "gray40",
  legend_title_size = 12,
  legend_position = "top",
  axis_margin = 10,
  axis_text_size = base_size,
  axis_title_font = subtitle_font,
  axis_title_size = 12,
  plot_margin = margin(30, 30, 30, 30)
)
}
\arguments{
\item{plot_margin}{}
}
\value{

}
\description{
transform any ggplot2 theme to an RTL theme with Persian fonts
}
