% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_farsh.R
\name{theme_farsh_fa}
\alias{theme_farsh_fa}
\title{Title
https://www.roozrang.com/\%d8\%b7\%d8\%b1\%d8\%a7\%d8\%ad\%db\%8c-\%d9\%87\%d9\%88\%db\%8c\%d8\%aa-\%d8\%a8\%d8\%b5\%d8\%b1\%db\%8c-\%d8\%af\%d8\%a7\%db\%8c\%d9\%86\%d8\%a7\%d9\%85\%db\%8c\%da\%a9-\%d9\%85\%d9\%88\%d8\%b2\%d9\%87-\%d9\%81\%d8\%b1\%d8\%b4-\%d8\%aa\%d9\%87\%d8\%b1\%d8\%a7/}
\usage{
theme_farsh_fa()
}
\value{

}
\description{
Title
https://www.roozrang.com/\%d8\%b7\%d8\%b1\%d8\%a7\%d8\%ad\%db\%8c-\%d9\%87\%d9\%88\%db\%8c\%d8\%aa-\%d8\%a8\%d8\%b5\%d8\%b1\%db\%8c-\%d8\%af\%d8\%a7\%db\%8c\%d9\%86\%d8\%a7\%d9\%85\%db\%8c\%da\%a9-\%d9\%85\%d9\%88\%d8\%b2\%d9\%87-\%d9\%81\%d8\%b1\%d8\%b4-\%d8\%aa\%d9\%87\%d8\%b1\%d8\%a7/
}
