% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hamshahri.R
\name{theme_hamshahri_modern}
\alias{theme_hamshahri_modern}
\title{Title}
\usage{
theme_hamshahri_modern(
  base_font = "A Rooznameh",
  base_size = 15.5,
  plot_title_font = "Zangar",
  plot_title_size = 30,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_font = base_font,
  subtitle_size = 20,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_font = base_font,
  strip_text_size = 17,
  strip_text_face = "plain",
  strip_text_margin = 10,
  caption_font = base_font,
  caption_size = 12,
  caption_face = "italic",
  caption_margin = 10,
  legend_font = "A Rooznameh",
  legend_font_size = 10,
  legend_color = "gray40",
  legend_title_size = 12,
  legend_position = "top",
  axis_text_size = base_size,
  axis_title_font = subtitle_font,
  axis_title_size = 12,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{legend_position}{}
}
\description{
Title
}
